Disable Failing Tests
All the data migration tests are not working at the moment.
The reason seems to be, that the files containing object model and databases are in different locations, when running from the terminal in contrast to running from within XCode.
As a solution I need to find out where the files are located while running from the terminal and either try both locations or have some way to detect which environment the test is currently running from.